feature[canny-source]: refactoring
